[김도형] 1. 쪽지보내기에서 골드구매 화면 연결 2. 하트보내기 기존 비용에서 10배씩 수정

+++ app/src/main/java/com/dohosoft/talkking/Data/UserData.java
@@ -92,7 +92,7 @@ public class UserData implements Serializable,Parcelable {
     public int Point;
     public int Grade;
 
-    public int ConnectDate;
+    public long ConnectDate;
     public long LastBoardWriteTime;
     public long LastAdsTime;
 
@@ -174,7 +174,7 @@ public class UserData implements Serializable,Parcelable {
         Point = in.readInt();
         Grade = in.readInt();
 
-        ConnectDate = in.readInt();
+        ConnectDate = in.readLong();
         LastBoardWriteTime = in.readLong();
         LastAdsTime = in.readLong();
 
@@ -267,7 +267,7 @@ public class UserData implements Serializable,Parcelable {
         parcel.writeInt(Point);
         parcel.writeInt(Grade);
 
-        parcel.writeInt(ConnectDate);
+        parcel.writeLong(ConnectDate);
         parcel.writeLong(LastBoardWriteTime);
         parcel.writeLong(LastAdsTime);
